utils: split helpers from bindings.c into utils.{c,h}
