+++ bindings.c
@@ -38,10 +38,11 @@
 #include <sys/vfs.h>
 
 #include "bindings.h"
+#include "config.h"
 #include "cgroups/cgroup.h"
 #include "cgroups/cgroup_utils.h"
 #include "memory_utils.h"
-#include "config.h"
+#include "utils.h"
 
 /* Define pivot_root() if missing from the C library */
 #ifndef HAVE_PIVOT_ROOT
@@ -580,87 +581,6 @@ static int is_dir(const char *path, int fd)
 	if (ret == 0 && S_ISDIR(statbuf.st_mode))
 		return 1;
 	return 0;
-}
-
-static int preserve_ns(const int pid, const char *ns)
-{
-	int ret;
-/* 5 /proc + 21 /int_as_str + 3 /ns + 20 /NS_NAME + 1 \0 */
-#define __NS_PATH_LEN 50
-	char path[__NS_PATH_LEN];
-
-	/* This way we can use this function to also check whether namespaces
-	 * are supported by the kernel by passing in the NULL or the empty
-	 * string.
-	 */
-	ret = snprintf(path, __NS_PATH_LEN, "/proc/%d/ns%s%s", pid,
-		       !ns || strcmp(ns, "") == 0 ? "" : "/",
-		       !ns || strcmp(ns, "") == 0 ? "" : ns);
-	if (ret < 0 || (size_t)ret >= __NS_PATH_LEN) {
-		errno = EFBIG;
-		return -1;
-	}
-
-	return open(path, O_RDONLY | O_CLOEXEC);
-}
-
-/**
- * in_same_namespace - Check whether two processes are in the same namespace.
- * @pid1 - PID of the first process.
- * @pid2 - PID of the second process.
- * @ns   - Name of the namespace to check. Must correspond to one of the names
- *         for the namespaces as shown in /proc/<pid/ns/
- *
- * If the two processes are not in the same namespace returns an fd to the
- * namespace of the second process identified by @pid2. If the two processes are
- * in the same namespace returns -EINVAL, -1 if an error occurred.
- */
-static int in_same_namespace(pid_t pid1, pid_t pid2, const char *ns)
-{
-	__do_close_prot_errno int ns_fd1 = -1, ns_fd2 = -1;
-	int ret = -1;
-	struct stat ns_st1, ns_st2;
-
-	ns_fd1 = preserve_ns(pid1, ns);
-	if (ns_fd1 < 0) {
-		/* The kernel does not support this namespace. This is not an
-		 * error.
-		 */
-		if (errno == ENOENT)
-			return -EINVAL;
-
-		return -1;
-	}
-
-	ns_fd2 = preserve_ns(pid2, ns);
-	if (ns_fd2 < 0)
-		return -1;
-
-	ret = fstat(ns_fd1, &ns_st1);
-	if (ret < 0)
-		return -1;
-
-	ret = fstat(ns_fd2, &ns_st2);
-	if (ret < 0)
-		return -1;
-
-	/* processes are in the same namespace */
-	if ((ns_st1.st_dev == ns_st2.st_dev) && (ns_st1.st_ino == ns_st2.st_ino))
-		return -EINVAL;
-
-	/* processes are in different namespaces */
-	return move_fd(ns_fd2);
-}
-
-static bool is_shared_pidns(pid_t pid)
-{
-	if (pid != 1)
-		return false;
-
-	if (in_same_namespace(pid, getpid(), "pid") == -EINVAL)
-		return true;
-
-	return false;
 }
 
 static bool write_string(const char *fnam, const char *string, int fd)
@@ -1348,34 +1268,6 @@ again:
 	return 0;
 }
 
-/*
- * append the given formatted string to *src.
- * src: a pointer to a char* in which to append the formatted string.
- * sz: the number of characters printed so far, minus trailing \0.
- * asz: the allocated size so far
- * format: string format. See printf for details.
- * ...: varargs. See printf for details.
- */
-static void must_strcat(char **src, size_t *sz, size_t *asz, const char *format, ...)
-{
-	char tmp[BUF_RESERVE_SIZE];
-	va_list		args;
-
-	va_start (args, format);
-	int tmplen = vsnprintf(tmp, BUF_RESERVE_SIZE, format, args);
-	va_end(args);
-
-	if (!*src || tmplen + *sz + 1 >= *asz) {
-		char *tmp;
-		do {
-			tmp = realloc(*src, *asz + BUF_RESERVE_SIZE);
-		} while (!tmp);
-		*src = tmp;
-		*asz += BUF_RESERVE_SIZE;
-	}
-	memcpy((*src) +*sz , tmp, tmplen+1); /* include the \0 */
-	*sz += tmplen;
-}
 
 /*
  * append pid to *src.
@@ -6011,19 +5903,6 @@ static bool permute_root(void)
 	return true;
 }
 
-static int preserve_mnt_ns(int pid)
-{
-	int ret;
-	size_t len = sizeof("/proc/") + 21 + sizeof("/ns/mnt");
-	char path[len];
-
-	ret = snprintf(path, len, "/proc/%d/ns/mnt", pid);
-	if (ret < 0 || (size_t)ret >= len)
-		return -1;
-
-	return open(path, O_RDONLY | O_CLOEXEC);
-}
-
 static bool cgfs_prepare_mounts(void)
 {
 	if (!mkdir_p(BASEDIR, 0700)) {
@@ -6041,7 +5920,7 @@ static bool cgfs_prepare_mounts(void)
 		return false;
 	}
 
-	cgroup_ops->mntns_fd = preserve_mnt_ns(getpid());
+	cgroup_ops->mntns_fd = preserve_ns(getpid(), "mnt");
 	if (cgroup_ops->mntns_fd < 0) {
 		lxcfs_error("Failed to preserve mount namespace: %s.\n", strerror(errno));
 		return false;
@@ -6105,7 +5984,7 @@ static void __attribute__((constructor)) lxcfs_init(void)
 		log_exit("Failed to initialize cgroup support");
 
 	/* Preserve initial namespace. */
-	init_ns = preserve_mnt_ns(getpid());
+	init_ns = preserve_ns(getpid(), "mnt");
 	if (init_ns < 0)
 		log_exit("Failed to preserve initial mount namespace");
